Don't force use of the credential_store value for DB password.
Currently, the ansible script forces use of the credential store's Nextcloud database password, even though currently this value is set within the playbook. If the value does not exist in user-defined values, a value is generated, placed in the credential store, and set for the playbook. 

This commit utilizes the playbook's variable instead of forcing use of the credential store's.